% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_analysis.R
\name{calculate_and_display_summary}
\alias{calculate_and_display_summary}
\title{Helps to calculate statistics and display summary}
\usage{
calculate_and_display_summary(merged_info)
}
\arguments{
\item{merged_info}{}
}
\description{
This method helps to calculate:
\enumerate{
\item minimum watch count
\item category or categories with minimum watch count
\item sum of duration in minutes of categories with the minimum watch count
\item total percentage of time spent on categories with the minimum watch count
\item maximum watch count
\item category or categories with maximum watch count
\item sum of duration in minutes of categories with the maximum watch count
\item total percentage of time spent on categories with the maximum watch count
}
}
