% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_analyse_top_videos.R
\name{calculate_proportion}
\alias{calculate_proportion}
\title{Helps to calculate proportion of views, count,comments, duration}
\usage{
calculate_proportion(filtered_video_details, full_video_detail)
}
\arguments{
\item{full_video_detail}{}
}
\value{
proportion_data
}
\description{
Helps to calculate proportion of views, count,comments, duration
}
\examples{
dummy_data <- data.frame(
videotitle = "Dummy Video",
duration = 10,
views = 1000,
likes = 500,
comments = 100,
points = 0
)
calculate_proportion(dummy_data, dummy_data)
}
