% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_comparator_plots.R
\name{create_bar_videos_plot}
\alias{create_bar_videos_plot}
\title{Bar plot 3: Number of videos}
\usage{
create_bar_videos_plot(channel_stats)
}
\arguments{
\item{channel_stats}{A data frame with the channel statistics.}
}
\value{
A bar plot.
}
\description{
Function to create a bar plot for the total number of videos of 2 or more channel IDs.
}
\examples{
# Providing a valid API Key and channel ids
api_key <- "AIzaSyB6df_K1PJy64w5VLZGYWyXSPYJ-TOoXVw"
channel_ids = c('UCtYLUTtgS3k1Fg4y5tAhLbw',
                'UCCezIgC97PvUuR4_gbFUs5g',
                'UCJQJAI7IjbLcpsjWdSzYz0Q')
# Call function to retrieve channel statistics data frame
channel_stats <- get_channel_stats(api_key, channel_ids)
bar_videos <- create_bar_videos_plot(channel_stats)
}
