% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_comparator_plots.R
\name{create_engagement_plot}
\alias{create_engagement_plot}
\title{Bar plot 4: Engagement level}
\usage{
create_engagement_plot(channel_stats)
}
\arguments{
\item{channel_stats}{A data frame with the channel statistics.}
}
\value{
A bar plot.
}
\description{
Function to create a barplot to provide a way to understand the relationship between the
number of subscribers and the number of views on the channel, helping to gauge the channel's
level of engagement with its audience.
}
\examples{
# Providing a valid API Key and channel ids
api_key <- "AIzaSyB6df_K1PJy64w5VLZGYWyXSPYJ-TOoXVw"
channel_ids = c('UCtYLUTtgS3k1Fg4y5tAhLbw',
                'UCCezIgC97PvUuR4_gbFUs5g',
                'UCJQJAI7IjbLcpsjWdSzYz0Q')
# Call function to retrieve channel statistics data frame
channel_stats <- get_channel_stats(api_key, channel_ids)
proportion_plot <- create_engagement_plot(channel_stats)
}
