% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_comparator_plots.R
\name{create_growth_plot}
\alias{create_growth_plot}
\title{Bubble graph 2: Channel growth comparator}
\usage{
create_growth_plot(channel_stats)
}
\arguments{
\item{channel_stats}{A data frame with the channel statistics.}
}
\value{
A bubble graph.
}
\description{
Creates a bubble plot that shows the relationship between the number of subscribers,
number of views and total videos per channel, per content creator adding the start date
of the channel.
}
\examples{
# Providing a valid API Key and channel ids
api_key <- "AIzaSyB6df_K1PJy64w5VLZGYWyXSPYJ-TOoXVw"
channel_ids = c('UCtYLUTtgS3k1Fg4y5tAhLbw', # Statquest
                'UCCezIgC97PvUuR4_gbFUs5g', # Corey Schafer
                'UCJQJAI7IjbLcpsjWdSzYz0Q') # Thu Vu
# Call function to retrieve channel statistics data frame
channel_stats <- get_channel_stats(api_key, channel_ids)
growth_plot <- create_growth_plot(channel_stats)
}
