% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_individual_channel_stats.R
\name{display_channel_info}
\alias{display_channel_info}
\title{Table with channel information}
\usage{
display_channel_info(channel_stats)
}
\arguments{
\item{channel_stats}{}
}
\value{
A table in the viewer.
}
\description{
Display the general information about a channel id.
}
\examples{
# Providing a valid API Key and channel ids
api_key <- "AIzaSyB6df_K1PJy64w5VLZGYWyXSPYJ-TOoXVw"
# Specifying a unique channel id
channel_id = c('UC2UXDak6o7rBm23k3Vv5dww') # Tina Huang
# Call function to retrieve channel statistics data frame
channel_stats <- get_channel_stats(api_key, channel_id)
# Call the function to display the table with the channel statistics
display_channel_info(channel_stats)
}
