% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_details_util.R
\name{Extract_Time_in_mins}
\alias{Extract_Time_in_mins}
\title{Helps to Extract Time in minutes
The YouTube time period follows ISO 8601 time that looks like PT1H45M35S
This method helps to extract time from above mentioned format to minutes
For instance:
PT1H30M30S is converted to 90.5 minutes using regex}
\usage{
Extract_Time_in_mins(input)
}
\arguments{
\item{input}{time in ISO 8601 format}
}
\value{
time in minutes
}
\description{
Helps to Extract Time in minutes
The YouTube time period follows ISO 8601 time that looks like PT1H45M35S
This method helps to extract time from above mentioned format to minutes
For instance:
PT1H30M30S is converted to 90.5 minutes using regex
}
\examples{
Extract_Time_in_mins("PT1H30M30S")
Extract_Time_in_mins("PT30M30S")
Extract_Time_in_mins("PT1H30S")
}
