% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YTExploreR_get_uploads.R
\name{format_date}
\alias{format_date}
\title{Date Formatter}
\usage{
format_date(month, year)
}
\arguments{
\item{month:}{two digit numeric value from 1-12, to indicate month. 01 for Jan, 02 for Feb,  etc.}

\item{year:}{four digit numeric value, to indicate year.}
}
\value{
Two formatted strings, one month apart:
Suitable to be plugged into the YouTube Activities API publishedBefore and publishedAfter parameters
}
\description{
Date Formatter
}
