% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channel_stats.R
\name{get_channel_stats}
\alias{get_channel_stats}
\title{Channel statistics}
\usage{
get_channel_stats(api_key, channel_ids)
}
\arguments{
\item{api_key}{An alphanumeric key to access the YouTube API.}

\item{channel_ids}{A character vector of channel identifiers used for retrieve statistics.}
}
\value{
A data frame with the statistics, including:
\itemize{
\item channelID: Unique identifier from the channel.
\item channelName: The channel's title.
\item numberSubscribers: The number of subscribers that the channel has.
\item numberViews: The number of times the channel has been viewed.
\item totalVideos: The number of public videos uploaded to the channel.
\item playlistId: The ID of the playlist that contains the channel's liked videos.
\item channelStartDate: The date and time that the channel was created.
\item country: The country with which the channel is associated.
\item thumbnails_present: Indicates whether thumbnails are present (Yes/No)
\item engagement_ratio: The ratio of subscribers to views (in percentage)
}
}
\description{
This function retrieves statistics for one or more YouTube channels using the YouTube API.
}
\examples{
# Set your API key
api_key <- "AIzaSyB6df_K1PJy64w5VLZGYWyXSPYJ-TOoXVw"
# Define channel IDs
channel_ids <- c("UCtYLUTtgS3k1Fg4y5tAhLbw",  # Statquest
                "UCLLw7jmFsvfIVaUFsLs8mlQ")  # Luke Barousse
# Call the function to get channel statistics
channel_stats <- get_channel_stats(api_key, channel_ids)
# Print the resulting data frame
print(channel_stats)
}
