% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YTExploreR_get_uploads.R
\name{get_monthly_uploads}
\alias{get_monthly_uploads}
\title{Monthly Upload Activity}
\usage{
get_monthly_uploads(chan_id, year)
}
\arguments{
\item{year}{}
}
\value{
sorted dataframe with columns associated with channel activity
}
\description{
Input a YouTube channel id and retrieves activities that are carried out by that channel.
This includes: uploading videos, creating playlists (of pre-existing videos), likes activity, social activity and more.
}
