% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_analyse_top_videos.R
\name{get_top10_videos}
\alias{get_top10_videos}
\title{Helps to co-ordinate all the function
Ranks and order the video details by calling get_video_and_rank
Then filters top 10 videos
Then helps to calculate proportion by calling calculate_proportion
Then prints and visualizes data}
\usage{
get_top10_videos(channelId)
}
\arguments{
\item{channelId}{}
}
\description{
Helps to co-ordinate all the function
Ranks and order the video details by calling get_video_and_rank
Then filters top 10 videos
Then helps to calculate proportion by calling calculate_proportion
Then prints and visualizes data
}
\examples{
get_top10_videos("UCtYLUTtgS3k1Fg4y5tAhLbw")
}
