% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_analyse_top_videos.R
\name{get_video_and_rank}
\alias{get_video_and_rank}
\title{Helps to rank the videos}
\usage{
get_video_and_rank(channelID, year)
}
\arguments{
\item{year}{}
}
\value{
videos a dataframe
}
\description{
This method:
gets video details from Get_Video_Detail method in video_details_util.R
then calculates ranks with a formula \code{(0.3*views) + (0.2 * likes) + (0.2 * comments)}
This formula is used because instead of giving equal weight age for views, likes and comments,
we want to emphasize on it's importance
}
\examples{
get_video_and_rank("UCtYLUTtgS3k1Fg4y5tAhLbw",2023)
}
