% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_category_util.R
\name{Get_Video_Category}
\alias{Get_Video_Category}
\title{Helps to Get Video Category details}
\usage{
Get_Video_Category(categoryid, api_key)
}
\arguments{
\item{categoryid}{It could be either single id or a comma seperated multiple ids}

\item{api_key}{API key to authenticate the access}
}
\value{
data frame with category details
}
\description{
This method uses video category api, which accepts single or multiple category ids,
to get all the category details.
The output dataframe contains:
\enumerate{
\item categoryId
\item categoryTitle
\item channelId
}
}
\examples{
Get_Video_Category("22,20", "AIzaSyBqrBJzAuitb-PpfyPrV7ABbLn8_nIbK3c")
Get_Video_Category("22", "AIzaSyBqrBJzAuitb-PpfyPrV7ABbLn8_nIbK3c")
}
