% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_details_util.R
\name{Get_Video_Detail}
\alias{Get_Video_Detail}
\title{Helps to get video details
All the essential details about a video or videos can be extracted from this method
This method supports single and multiple videos.
This method gives details such as:
\enumerate{
\item channelId
\item videoID
\item video Title
\item published At
\item publishedmonth
\item categoryid
\item duration
\item definition
\item views
10.likes
11.favourites
12.comments
13.publishedyear
and converts data into appropriate type
}}
\usage{
Get_Video_Detail(video_id, api_key)
}
\arguments{
\item{video_id}{id of a video or comma seperated ids of a video}

\item{api_key}{api key to authenticate API}
}
\value{
data frame with details about the video
}
\description{
Helps to get video details
All the essential details about a video or videos can be extracted from this method
This method supports single and multiple videos.
This method gives details such as:
\enumerate{
\item channelId
\item videoID
\item video Title
\item published At
\item publishedmonth
\item categoryid
\item duration
\item definition
\item views
10.likes
11.favourites
12.comments
13.publishedyear
and converts data into appropriate type
}
}
\examples{
Get_Video_Detail("Ks-_Mh1QhMc", "AIzaSyBqrBJzAuitb-PpfyPrV7ABbLn8_nIbK3c")
Get_Video_Detail("Ks-_Mh1QhMc,Ks-_Mh1QhMc", "AIzaSyBqrBJzAuitb-PpfyPrV7ABbLn8_nIbK3c")
}
