% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video_ids.R
\name{get_video_ids}
\alias{get_video_ids}
\title{Auxiliar function to get the video identifiers from the YouTube API.
The playlist id information can be taken from the channel statistics data frame retrieved by the get_channel_stats function.
This facilitates the level access for the analysis:
(user: channel_id -> get_channel_stats -> playlist_id -> video_ids -> video_details -> channel_uploads -> etc.)}
\usage{
get_video_ids()
}
\arguments{
\item{api_key}{Is the alphanumeric key to get access to the YouTube API.}

\item{playlist_id}{The ID that YouTube uses to uniquely identify the playlist.}
}
\value{
A vector with one or more video ids.
}
\description{
Auxiliar function to get the video identifiers from the YouTube API.
The playlist id information can be taken from the channel statistics data frame retrieved by the get_channel_stats function.
This facilitates the level access for the analysis:
(user: channel_id -> get_channel_stats -> playlist_id -> video_ids -> video_details -> channel_uploads -> etc.)
}
