% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_analysis.R
\name{preprare_data}
\alias{preprare_data}
\title{Helps to prepare data for preference breakdown}
\usage{
preprare_data(video_ids)
}
\arguments{
\item{video_ids}{}
}
\value{
merged_info(data frame)
}
\description{
This method gets the data from Get_Video_Detail method in video_details_util.R
Then groups the data by category and channelID
Then calculates count and sum of duration for each category and channel
Then calculates the percentage of time spent in each category
After preparing the above data,
calls Get_Video_Category method in video_category_util.R to get category details
Now merges the title from category details to the data frame in which we calculated
}
\examples{
preprare_data("ZTt9gsGcdDo,Qf06XDYXCXI")
}
