% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_analyse_top_videos.R
\name{print_videos}
\alias{print_videos}
\title{Helps to print top 10 videos}
\usage{
print_videos(filtered_video_details, reverse = TRUE)
}
\arguments{
\item{reverse}{}
}
\description{
This method helps to rename the videos and helps to print it
}
\examples{
dummy_data <- data.frame(
videotitle = "Dummy Video",
duration = 10,
views = 1000,
likes = 500,
comments = 100,
points = 0
)
print_videos(dummy_data)
}
