% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_analysis.R
\name{tabulate_overallsummary}
\alias{tabulate_overallsummary}
\title{Tabulates the overall summary on category}
\usage{
tabulate_overallsummary(merged_info)
}
\arguments{
\item{merged_info}{}
}
\description{
Tabulates the overall summary on category
}
