% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_analysis.R
\name{tabulate_summary}
\alias{tabulate_summary}
\title{Tabulates summary on least and most favorite category}
\usage{
tabulate_summary(merged_info)
}
\arguments{
\item{merged_info}{}
}
\value{
: gives table
}
\description{
Tabulates summary on least and most favorite category
}
