% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_analysis.R
\name{visualize}
\alias{visualize}
\title{Helps to display category in pi chart}
\usage{
visualize(merged_info)
}
\description{
This method helps to create pi chart
with each slice representing the count of each category
}
\examples{
merged_data <- data.frame(categoryid = c("22", "27"),
      channelId = c("UCAuUUnT6oDeKwE6v1NGQxug", "UCtYLUTtgS3k1Fg4y5tAhLbw"),
      categoryTitle = c("Entertainment","Education"),
      count = c(1, 19),
      duration_sum = c(21.05, 399.95),
      percentage_duration = c(5, 95))
visualize(merged_data)
}
