% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_analyse_top_videos.R
\name{visualize_proportion}
\alias{visualize_proportion}
\title{Helps to visualize the proportion}
\usage{
visualize_proportion(proportion_data)
}
\arguments{
\item{proportion_data}{}
}
\description{
Helps to visualize the proportion
}
\examples{
dummy_data <- data.frame(
Category = c("Views", "Likes", "Comments", "Duration"),
Proportion = c(0.25, 0.35, 0.20, 0.30) # Example proportions (you can change these values as needed)
)
visualize_proportion(dummy_data)
}
