% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YTExploreR-package.R
\docType{package}
\name{YTExploreR-package}
\alias{YTExploreR}
\alias{YTExploreR-package}
\title{YTExploreR: What the Package Does (One Line, Title Case)}
\description{
Helps youtube watchers and content creator to get more insights on youtube videos, channel
}
\author{
\strong{Maintainer}: Karthiga Sethu Sethuramalingam \email{karthigaksk@gmail.com}

Authors:
\itemize{
  \item Nayeli Montiel \email{naye.montiel.1806@gmail.com}
  \item Andrew Sarracini \email{andrewsarracini1@gmail.com}
}

}
\keyword{internal}
